% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{earthcon}
\alias{earthcon}
\title{Sets Earth Constants as global variables.}
\usage{
earthcon(ai, bi)
}
\arguments{
\item{ai}{no idea}

\item{bi}{no idea}
}
\description{
Sets Earth Constants as global variables.
}

