% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{gd2gc}
\alias{gd2gc}
\title{geodetic latitude to geocentric latitude}
\usage{
gd2gc(flatgdi, altkmi)
}
\arguments{
\item{flatgdi}{geodetic latitude deg}

\item{altkmi}{altitide in km}
}
\value{
geocentric latitude in deg
}
\description{
geodetic latitude to geocentric latitude
}

