% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lla2xyz.R
\name{lla2xyz}
\alias{lla2xyz}
\title{Convert LLA to ECEF}
\usage{
lla2xyz(vec)
}
\arguments{
\item{vec}{lat)deg), long(deg), and altitude(km)}
}
\value{
A 3-vec containing xyz for ECEF (km)
}
\description{
Convert LLA to ECEF
}
\details{
This converter uses N latitude and E longitude. Heights are in meters, and are ellipsoidal heights. There is no geoid model included.
The WGS 84 ellipsoid is used.
}

