% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llh2xyz.R
\name{llh2xyz}
\alias{llh2xyz}
\title{Title}
\usage{
llh2xyz(vec)
}
\arguments{
\item{altkmi}{}
}
\description{
Title
}
\details{
This converter uses N latitude and E longitude. Heights are in meters, and are ellipsoidal heights. There is no geoid model included.
The WGS 84 ellipsoid is used.
}

