% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{radcur}
\alias{radcur}
\title{compute the radii at the geodetic latitude lat (in degrees)}
\usage{
radcur(lati)
}
\arguments{
\item{lati}{geodetic latitude in degrees}
}
\value{
r, rn, rm in km
}
\description{
compute the radii at the geodetic latitude lat (in degrees)
}

