% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{rearth}
\alias{rearth}
\title{physical radius of earth from geodetic latitude}
\usage{
rearth(lati)
}
\arguments{
\item{lati}{}
}
\value{
radius of earth at lat
}
\description{
physical radius of earth from geodetic latitude
}

