% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{earthcon}
\alias{earthcon}
\title{Sets Earth Constants as global variables.}
\usage{
earthcon(a, b)
}
\arguments{
\item{a}{no idea}

\item{b}{no idea}
}
\description{
Sets Earth Constants as global variables.
}

