% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gc2gd}
\alias{gc2gd}
\title{geocentric latitude to geodetic latitude}
\usage{
gc2gd(flatgci, altkmi)
}
\arguments{
\item{flatgci}{geocentric latitude deg}

\item{altkmi}{altitide in km}
}
\value{
geodetic latitude in deg
}
\description{
geocentric latitude to geodetic latitude
}

