% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gd2gc}
\alias{gd2gc}
\title{geodetic latitude to geocentric latitude}
\usage{
gd2gc(flatgd, altkm)
}
\arguments{
\item{flatgd}{geodetic latitude deg}

\item{altkm}{altitide in km}
}
\value{
geocentric latitude in deg
}
\description{
geodetic latitude to geocentric latitude
}

