% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{llenu}
\alias{llenu}
\title{latitude longitude to east,north,up unit vectors}
\usage{
llenu(flati, floni)
}
\arguments{
\item{flati}{latitude in degees N}

\item{floni}{longitude in degrees E}
}
\value{
enu[3[3]]  packed 3-unit vectors / each a 3 vector
}
\description{
latitude longitude to east,north,up unit vectors
}

