% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz2lla.R
\name{xyz2lla}
\alias{xyz2lla}
\title{ECEF Coordinates to Lat, Long, Altitude}
\usage{
xyz2lla(vec)
}
\arguments{
\item{xvec}{xyz ECEF location}
}
\value{
A 3-vec containing lat(deg), long(deg), altitude(km)
}
\description{
ECEF Coordinates to Lat, Long, Altitude
}

